requirements.txt
instagrapi==2.1.3
pillow==11.1.0
openpyxl==3.1.2  
pandas==2.2.0 
python-dotenv==1.0.1
pydub==0.25.1